Extension work. This fixes a bunch of issues with the application crashing pretty hard when extensions would just disappear. Not, this handles extensions disappearing gracefully, by removing all reference to the sockets and setting the mappings' Ready flag to false. In addition, a little refactoring was done to arrange Extension methods logically, and separate tasks into different classes.
Also, since there is only ever one extension system in the application, ExtensionSystem is now akin to a Java Singleton. Call ExtensionSystem.Instance to get the running instance, instead of creating new ones, or whatever. Not that anyone will be doing that, since it's in the application itself, but still. It's there.