Starting refactor into a service system instead of just a chat network system. Should allow more flexibility, but maaaan are there going to be a bunch of changes coming here. This commit lists the major refactors of changing "networks" to "services" and renaming directories, internal references, as well as some cleanup.
This also refactors NetworkBot into ServiceWrapper. Finally, the name of this thing makes sense. It's just that- a wrapper for a service, providing a generic container for one.

Messages got another overhaul- this refactor removes the enum-byte "MessageType" and replaces it with a single boolean (maybe two)- is_private. This is generic in Chat.Message, though- another change here is that services call a basic method from it's parent, but can pass along extra information with the event. For example, you can now tell if an IRC message is an action or not, because it passes along "is_action" inside of "message" in the "message_recieved" event.

In the chat service's generic User object, UniqueName is changed to UniqueID instead.