Oh god, the changelist for this one..
- ServiceWrapper (formerly NetworkWrapper) is gone. Services are directly handled in the Core.ConnectedServices dictionary.
- Configuration is pushed into a small helper class, and checking is moved out of Core.
- Identifiers table in the configuration database is renamed to Locations.
- Ripped out a bunch of unused delegates and events from the chat and service framework. Unused.

- New class, ServiceHelper:
Designed to load in and manage running services by working at low-level with the application DLLs and such
Handles the creation of ServiceConnector objects. Much faster this time, thanks to the IServiceStarter interface.
Said interface is to be implemented in a base class (no namespace) in the service DLLs. This class is loaded directly, the interface wrapped, and the method CreateServiceInstance is called.
In this method, developers are 100% free to do whatever they need to do to get a service connector working and set up, before it's passed back to the main application.
After the service connector is passed back, it's checked for validity and then the service's Setup method is called. This method is passed the ROOT configuration directory of Suibhne. Use it however you need to.

- Service configuration changes.
In the setup method, the service is now passed the root configuration directory, rather than the service's config root. This is so services can dig into the other directories easier, check for dependencies in a shared folder, etc.
In addition, it's no longer required that the service.ini file has a [Service] section with what type it is. Right now there's no support in the GUI, but it's been moved into the database's Locations table.

- Event changes
Event routing is done by the EventManager. It's first looked at by the internal, defined event handlers; stuff like message_recieved and such. These event handlers can pass on the events to where they need to go, but if one
isn't defined for an event, it's checked against the list of subscribed extensions' events. If it's STILL not found, the system throws out an error into the logs.