[feature/refactor] - Modified the event code to change how routing works. It's now intended that events only route events to the service connector; the connector attaches additional data to the routing object so it can handle the response later. The "routing" object is passed around and kept the same throughout the event managing lifestyle.
In addition to this, the Chat service type's Message no longer has a location ID hooked to it. Feel free to add a variable back in if you need it in your processing, but that's now up to the service implementation.

This update also fixes the Help command. The system will now look for an event, "command_help_response", to get the information back.