[major/refactor] - Removing a lot of stuff related to access levels being coded in. Planned that it will still go off a byte, but service connectors will have less of an immediate say in how access levels are assigned- there's only going to be one place to look, and that's in the database. If a service has a say in the level, it's to pass along a 'recommended' level along with the user object in the event. The system can evaluate it from there. As of now, nothing looks at access levels, because nothing NEEDS to look at them.
Refactored LocationManager into ServiceManager, and ServiceHelper into Services.ServiceLoader.

Events Handlers/InformationRequest:
Refactored information_request's location request into service instead, to lock in changes that locations aren't a thing anymore. Again, serviceID is now used here. (For this event, the request asks for a serviceID in the root event)

ExtensionSystem:
Fixed a few things in the core extension system so that the extension processes start in the background, redirecting output to the command line instead of spamming windows.

Extension:
Removed "is_private" check from Extension.cs- as long as that's a part of the routing object, the service should be able to route it fine! No need to put it in the event again, that should've been handled already.

Chat Service:
Removed "Listened" bits from ChatService. Again, locations are for services to handle now, not for the base system. Some chat services don't have a concept of one user, many locations.

User:
Again, removing access level stuff from here. No longer in the user objects themselves, in the database.

Service Base:
HandleConnectionComplete changed to ConnectionComplete.
HandleDiconnectComplete changed to DisconnectionComplete, and implemented with the OnDisconnectComplete event.